First progress update
Work was done locally before this commit. Still need to finalize UI for card results and then do a lot of the backend stuff. Ignored the contents of the card art folder. While important for the DB, art is (or will be) downloaded almost every time the database is booted. Included libraries alongside for now. May ignore in future commits.